requirements.txt
# Plant Disease Detection System - Production Requirements

# Core ML libraries (CPU optimized for deployment)
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html

# Image processing
Pillow==10.1.0
numpy==1.24.3

# Data processing
pandas==2.0.3
scikit-learn==1.3.0

# Web application
streamlit==1.28.1
plotly==5.17.0

# Progress bars and utilities
tqdm==4.66.1

# Production optimizations
gunicorn==21.2.0
watchdog==3.0.0

# Error handling and logging
loguru==0.7.2

# File handling
pathlib2==2.3.7